Fixed type for make_id task functions.

+++ src/TiledArray/reduce_task.h
@@ -76,7 +76,7 @@ namespace TiledArray {
       madness::Future<result_type> result_;
       madness::Spinlock lock_;
 
-      virtual void get_id(std::pair<void*,unsigned long>& id) const {
+      virtual void get_id(std::pair<void*,unsigned short>& id) const {
           return madness::PoolTaskInterface::make_id(id, *this);
       }
 
@@ -306,7 +306,7 @@ namespace TiledArray {
 
         const second_argument_type& right() const { return right_.get(); }
 
-        virtual void get_id(std::pair<void*,unsigned long>& id) const {
+        virtual void get_id(std::pair<void*,unsigned short>& id) const {
             return madness::PoolTaskInterface::make_id(id, *this);
         }
